refactor(dataloader): rename DataLoader to DataLoaderClient
change method to a void
